more warnings

+++ src/maths/codegen/frontend.h
@@ -46,14 +46,14 @@
 #define ORT_MATH_FRONTEND_REDUCTION_AXIS_OP_DECL(op, type)            \
     void ORT_MATH_FRONTEND_OP_AXIS_SYMBOL(op, type) (                 \
         void* result, const void* a,                                  \
-        const size_t *input_shape, size_t input_dims,                 \
-        const size_t *output_shape, size_t output_dims,               \
+        const int64_t *input_shape, size_t input_dims,                 \
+        const int64_t *output_shape, size_t output_dims,               \
         size_t axis)
 #define ORT_MATH_FRONTEND_SERIAL_REDUCTION_AXIS_OP_DECL(op, type)     \
     void ORT_MATH_FRONTEND_OP_AXIS_SYMBOL(serial_##op, type) (        \
         void* result, const void* a,                                  \
-        const size_t *input_shape, size_t input_dims,                 \
-        const size_t *output_shape, size_t output_dims,               \
+        const int64_t *input_shape, size_t input_dims,                 \
+        const int64_t *output_shape, size_t output_dims,               \
         size_t axis)
 
 #define ORT_MATH_FRONTEND_BINARY_OP_IMPL(op_name, c_type, operator)   \
